Transactions for table management

+++ ISA/src/main/java/ISAProject/controller/OrderItemController.java
@@ -125,8 +125,14 @@ public class OrderItemController {
             value = "/removeOrderItem/{id}",
             method = RequestMethod.DELETE)
     public ResponseEntity<OrderItem> removeOrderItem(@PathVariable("id") Long id) {
-        orderItemService.delete(id);
-        return new ResponseEntity<OrderItem>(HttpStatus.NO_CONTENT);
+        OrderItem orderItem = orderItemService.findOne(id);
+        if (orderItem.getOiStatus().equals("Waiting for waiter") || orderItem.getOiStatus().equals("Waiting")) {
+            orderItemService.delete(id);
+            return new ResponseEntity<OrderItem>(HttpStatus.NO_CONTENT);
+        }
+        else {
+            return new ResponseEntity<OrderItem>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @MessageMapping("/updateOrderItem/{rid}")
